requirements.txt
# ── Backend ────────────────────────────────────────────────────────────────── #
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
sqlalchemy>=2.0.30
aiosqlite>=0.20.0
python-multipart>=0.0.9

# ── Dashboard ──────────────────────────────────────────────────────────────── #
streamlit>=1.35.0
plotly>=5.22.0
requests>=2.32.0
pandas>=2.2.0

# ── Simulation (Isaac Lab / Isaac Sim managed separately via conda env) ───── #
# These are present in the env_isaaclab conda environment:
#   isaaclab, isaaclab_assets, gymnasium, torch, numpy
# Install extras via: isaaclab.bat -p -m pip install <package>

# ── Testing ────────────────────────────────────────────────────────────────── #
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# ── Dev utilities ──────────────────────────────────────────────────────────── #
python-dotenv>=1.0.0
pyyaml>=6.0.1
